whisper_base/src/requirements.txt
openai-whisper
soundfile
scipy